refactor compression-related files & use indexes
revisiting outdated code before introducing vq compression
