+++ src/utils/data/index.ts
@@ -1,2 +1,4 @@
 export { default as bufferToObjectUrl } from './bufferToObjectUrl';
 export { default as objectUrlToBuffer } from './objectUrlToBuffer';
+export { default as decompressLzssBuffer } from './decompressLzssBuffer';
+export { default as compressLzssBuffer } from './compressLzssBuffer';
