+++ src/utils/textures/index.ts
@@ -1,3 +1,7 @@
 export * from './RgbaColor';
 export * from './TextureSize';
 export * from './TextureColorFormat';
+export * from './SourceTextureData';
+export { default as uvToCssPathPoint } from './uvToCssPathPoint';
+export * from './serialize';
+export * from './parse';
