+++ src/utils/textures/parse/index.ts
@@ -1,6 +1,5 @@
-export { default as compressTextureBuffer } from './compressTextureBuffer';
-export { default as decompressTextureBuffer } from './decompressTextureBuffer';
 export { default as encodeZMortonPosition } from './encodeZMortonPosition';
 export { default as getTextureColorFormat } from './getTextureColorFormat';
 export { default as getTextureSize } from './getTextureSize';
 export { default as getTextureWrappingFlags } from './getTextureWrappingFlags';
+export { default as loadTextureFile } from './loadTextureFile';
