+++ src/utils/textures/serialize/index.ts
@@ -1,0 +1,1 @@
+export { default as decodeZMortonPosition } from './decodeZMortonPosition';
